Create service with firebase
